+++ DemoClient.cpp
@@ -1,4 +1,10 @@
-
+//============================================================================
+// Name        : DemoClient.cpp
+// Author      : 
+// Version     :
+// Copyright   : Your copyright notice
+// Description : Hello World in C++, Ansi-style
+//============================================================================
 
 #include <iostream>
 #include <string.h>
@@ -9,157 +15,96 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <netdb.h>
+#include <fstream>
 
 using namespace std;
 
-int main()
-{
-    /* ---------- INITIALIZING VARIABLES ---------- */
+// Hàm cắt file và trả về 1 nếu là file được nhận, 0 nếu là đoạn thông điệp chat
+int sub(string *a);
 
-    /*  
-       1. client is a file descriptor to store the values 
-       returned by the socket system call and the accept 
-       system call.
-       2. portNum is for storing port number on which
-       the accepts connections
-       3. isExit is bool variable which will be used to 
-       end the loop
-       4. The client reads characters from the socket 
-       connection into a dynamic variable (buffer).
-       5. A sockaddr_in is a structure containing an internet 
-       address. This structure is already defined in netinet/in.h, so
-       we don't need to declare it again.
-        DEFINITION:
-        struct sockaddr_in
-        {
-          short   sin_family;
-          u_short sin_port;
-          struct  in_addr sin_addr;
-          char    sin_zero[8];
-        };
-        6. serv_addr will contain the address of the server
-    */
-
-    int client;
-    int portNum = 1220; // NOTE that the port number is same for both client and server
-    bool isExit = false;
-    int bufsize = 1024;
-    char buffer[bufsize];
-    char* ip = "127.0.0.1";
-
-    struct sockaddr_in server_addr;
-
-    client = socket(AF_INET, SOCK_STREAM, 0);
-
-    /* ---------- ESTABLISHING SOCKET CONNECTION ----------*/
-    /* --------------- socket() function ------------------*/
-
-    if (client < 0) 
-    {
-        cout << "\nError establishing socket..." << endl;
-        exit(1);
-    }
-
-    /*
-        The socket() function creates a new socket.
-        It takes 3 arguments,
-            a. AF_INET: address domain of the socket.
-            b. SOCK_STREAM: Type of socket. a stream socket in 
-            which characters are read in a continuous stream (TCP)
-            c. Third is a protocol argument: should always be 0. The 
-            OS will choose the most appropiate protocol.
-            This will return a small integer and is used for all 
-            references to this socket. If the socket call fails, 
-            it returns -1.
-    */
-
-    cout << "\n=> Socket client has been created..." << endl;
-    
-    /* 
-        The variable serv_addr is a structure of sockaddr_in. 
-        sin_family contains a code for the address family. 
-        It should always be set to AF_INET.
-        htons() converts the port number from host byte order 
-        to a port number in network byte order.
-    */
-
-    server_addr.sin_family = AF_INET;
-    server_addr.sin_port = htons(portNum);
-
-    // this function returns returns 1 if the IP is valid
-    // and 0 if invalid
-    // inet_pton converts IP to packets
-    // inet_ntoa converts back packets to IP
-    //inet_pton(AF_INET, ip, &server_addr.sin_addr);
-
-    /*if (connect(client,(struct sockaddr *)&server_addr, sizeof(server_addr)) == 0)
-        cout << "=> Connection to the server " << inet_ntoa(server_addr.sin_addr) << " with port number: " << portNum << endl;*/
+int main() {
+	int client, server;
+	int portNum = 1700;
+	bool isExit = false;
+	int bufsize = 1024;
+	char buffer[bufsize];
+	char *ip = "192.168.11.17";
+	ofstream file("input.txt");
 
 
-    /* ---------- CONNECTING THE SOCKET ---------- */
-    /* ---------------- connect() ---------------- */
+	struct sockaddr_in server_addr;
 
-    if (connect(client,(struct sockaddr *)&server_addr, sizeof(server_addr)) == 0)
-        cout << "=> Connection to the server port number: " << portNum << endl;
+	//init socket
+	client = socket(AF_INET, SOCK_STREAM, 0);
+	if (client < 0) {
+		cout << "Error creating socket..." << endl;
+		exit(1);
+	}
 
-    /* 
-        The connect function is called by the client to 
-        establish a connection to the server. It takes 
-        three arguments, the socket file descriptor, the 
-        address of the host to which it wants to connect 
-        (including the port number), and the size of this 
-        address. 
-        This function returns 0 on success and -1 
-        if it fails.
-        Note that the client needs to know the port number of
-        the server but not its own port number.
-    */
+	cout << "Client socket created" << endl;
 
-    cout << "=> Awaiting confirmation from the server..." << endl; //line 40
-    recv(client, buffer, bufsize, 0);
-    cout << "=> Connection confirmed, you are good to go...";
+	server_addr.sin_family = AF_INET;
+	server_addr.sin_port = htons(portNum);
+	server_addr.sin_addr.s_addr = inet_addr("192.168.11.93");
 
-    cout << "\n\n=> Enter # to end the connection\n" << endl;
+	//connecting socket server
+	if (connect(client, (struct sockaddr*) &server_addr, sizeof(server_addr))
+			== 0) {
+		cout << "Connecting to server..." << endl;
+	}
 
-    // Once it reaches here, the client can send a message first.
+	recv(client, buffer, bufsize, 0);
+	cout << "Connection confirmed" << endl;
 
-    do {
-        cout << "Client: ";
-        do {
-            cin >> buffer;
-            send(client, buffer, bufsize, 0);
-            if (*buffer == '#') {
-                send(client, buffer, bufsize, 0);
-                *buffer = '*';
-                isExit = true;
-            }
-        } while (*buffer != 42);
+	cout << "Enter # to end the connection." << endl;
 
-        cout << "Server: ";
-        do {
-            recv(client, buffer, bufsize, 0);
-            cout << buffer << " ";
-            if (*buffer == '#') {
-                *buffer = '*';
-                isExit = true;
-            }
+	//communicate client - server
+	do {
+		cout << "Client: ";
+		do {
+			cin >> buffer;
+			send(client, buffer, bufsize, 0);
+			if (*buffer == '#') {
+				send(client, buffer, bufsize, 0);
+				*buffer = '*';
+				isExit = true;
+			}
+		} while (*buffer != 42);
 
-        } while (*buffer != 42);
-        cout << endl;
+		cout << "Server: ";
+		do {
+			recv(client, buffer, bufsize, 0);
+			memset(buffer, 0, sizeof(buffer));
+			string buff(buffer);
+			int fi = sub(&buff);
+//			cout << buffer << " ";
+			if(fi == 1){
+				fstream f;
+				f.open("input.txt", ios::in | ios::out);
+				f << buff;
+			}
+			else cout << buff << " ";
+			if (buff == "#") {
+				*buffer = '*';
+				isExit = true;
+			}
+		} while (*buffer != 42);
+		cout << endl;
+	} while (!isExit);
 
-    } while (!isExit);
+	cout << "Connection terminaler..." << endl;
+	cout << "Goodbye.." << endl;
 
-    /* ---------------- CLOSE CALL ------------- */
-    /* ----------------- close() --------------- */
+	close(client);
 
-    /* 
-        Once the server presses # to end the connection,
-        the loop will break and it will close the server 
-        socket connection and the client connection.
-    */
+	return 0;
+}
 
-    cout << "\n=> Connection terminated.\nGoodbye...\n";
-
-    close(client);
-    return 0;
+int sub(string *a) {
+	string b = a->substr(0, 2);
+	*a = a->erase(0, 2);
+	if (b == "fi") {
+		return 1;
+	} else
+		return 0;
 }
